+++ examples/kyb-app/src/pages/CollectionFlow/components/organisms/KYBView/views/CompanyDocumentsView/company-documents.ui-schema.ts
@@ -3,7 +3,7 @@ import { UiSchema } from '@rjsf/utils';
 export const companyDocumetsUISchema: UiSchema = {
   'ui:options': {
     submitButtonOptions: {
-      submitText: 'Continue',
+      submitText: 'Submit',
     },
   },
   registrationCertificate: {