+++ packages/ui/src/components/organisms/DynamicForm/components/layouts/buttons/SubmitLayout/SubmitLayout.tsx
@@ -1,23 +1,25 @@
-import { Button } from '@components/atoms';
+import { SubmitButton } from '@components/molecules/SubmitButton';
 import { SubmitButtonProps, getSubmitButtonOptions } from '@rjsf/utils';
 import clsx from 'clsx';
 
 export const SubmitLayout = ({ uiSchema }: SubmitButtonProps) => {
   const { norender, submitText, props } = getSubmitButtonOptions(uiSchema);
   const disabled = Boolean(uiSchema['ui:options']?.submitButtonOptions?.props?.disabled);
+  //@ts-ignore
+  const isLoading = uiSchema['ui:options']?.submitButtonOptions?.isLoading as boolean;
 
   if (norender) return null;
 
   return (
     <div className={clsx('flex justify-end', props.layoutClassName as string)}>
-      <Button
-        // TO DO: checkout why variant isnt working as intended
+      <SubmitButton
         className="bg-secondary text-secondary-foreground hover:bg-secondary/80 shadow-sm"
-        type="submit"
+        loaderClassName="text-secondary-foreground"
         disabled={disabled}
+        isLoading={isLoading}
       >
         {submitText}
-      </Button>
+      </SubmitButton>
     </div>
   );
 };