+++ services/workflows-service/src/workflow/types/params.ts
@@ -1,3 +1,8 @@
+export interface FindLastActiveFlowParams {
+  workflowDefinitionId: string;
+  businessId: string;
+}
+
 export interface GetLastActiveFlowParams {
   email: string;
   workflowRuntimeDefinitionId: string;