+++ services/workflows-service/src/workflow/workflow-runtime-data.repository.intg.test.ts
@@ -19,6 +19,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import { PrismaService } from '@/prisma/prisma.service';
 import { EntityRepository } from '@/common/entity/entity.repository';
 import { ProjectScopeService } from '@/project/project-scope.service';
+import { EndUserService } from '@/end-user/end-user.service';
 
 describe('#Workflow Runtime Repository Integration Tests', () => {
   let workflowRuntimeRepository: WorkflowRuntimeDataRepository;
@@ -29,6 +30,7 @@ describe('#Workflow Runtime Repository Integration Tests', () => {
   beforeAll(async () => {
     const servicesProviders = [
       EndUserRepository,
+      EndUserService,
       FilterService,
       FilterRepository,
       ProjectScopeService,