+++ services/workflows-service/src/workflow/workflow.controller.external.ts
@@ -28,6 +28,8 @@ import { Public } from '@/common/decorators/public.decorator';
 import { WorkflowHookQuery } from '@/workflow/dtos/workflow-hook-query';
 import { HookCallbackHandlerService } from '@/workflow/hook-callback-handler.service';
 import { UseCustomerAuthGuard } from '@/common/decorators/use-customer-auth-guard.decorator';
+import { GetActiveFlowDto } from '@/workflow/dtos/get-active-workflow-input.dto';
+import { UseKeyAuthOrSessionGuard } from '@/common/decorators/use-key-auth-or-session-guard.decorator';
 
 @swagger.ApiBearerAuth()
 @swagger.ApiTags('external/workflows')
@@ -63,6 +65,19 @@ export class WorkflowControllerExternal {
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
   async getWorkflowDefinition(@common.Param() params: WorkflowDefinitionWhereUniqueInput) {
     return await this.service.getWorkflowDefinitionById(params.id);
+  }
+
+  @common.Get('/active-flow')
+  @UseKeyAuthOrSessionGuard()
+  async getActiveFlow(@common.Query() query: GetActiveFlowDto) {
+    const activeWorkflow = await this.service.getLastActiveFlow({
+      email: query.email,
+      workflowRuntimeDefinitionId: query.workflowRuntimeDefinitionId,
+    });
+
+    return {
+      result: activeWorkflow,
+    };
   }
 
   @common.Get('/:id')